Update the hot spring water function
When I just looked at the logic again, I realized that the springwater rush part is redundant. For the rush, you need access to "Lower Kakariko Village". But if you have access to that room and got a bottle, which you need anyways, the Elde Inn water becomes logically reachable automatically.